[TASK] Refactor xapi-publisher
split project into separate modules
